#1 - create views and consume endpoints

+++ src/components/Favorite.vue
@@ -2,10 +2,17 @@
   <div class="container">
     <div class="row">
       <div class="col-sm-10">
-        <h1>Favorite Things</h1>
+        <br><br>
+          <button type="button" class="btn btn-success btn-sm" v-b-modal.favorite-modal>Add Favorite Thing</button>
+
         <hr><br><br>
         <alert :message="message" v-if="showMessage"></alert>
-        <button type="button" class="btn btn-success btn-sm" v-b-modal.favorite-modal>Add Favorite Thing</button>
+        <div class="linkcontainer">
+           <router-link :to="{ name: 'Favorite' }">Favorites</router-link>
+          <router-link :to="{ name: 'Category' }">Categories</router-link>
+        </div>
+
+
         <br><br>
         <table class="table table-hover">
           <thead>
@@ -14,7 +21,7 @@
               <th scope="col">Category</th>
               <th scope="col">Description</th>
               <th scope="col">created date</th>
-              <th scope="col">meta data</th>
+              <th scope="col">more info</th>
               <th scope="col">rank</th>
               <th></th>
             </tr>
@@ -24,7 +31,7 @@
               <td>{{favorite.title}}</td>
               <td>{{favorite.category.type}}</td>
               <td>{{favorite.description}}</td>
-              <td>{{favorite.createdAt}}</td>
+              <td>{{formatDate(favorite.createdAt)}}</td>
               <td>{{favorite.metaData}}</td>
                <td>{{favorite.rank}}</td>
               <td>
@@ -60,7 +67,6 @@
           <b-form-input id="form-description-input"
                         type="text"
                         v-model="addFavoriteForm.description"
-                        required
                         placeholder="Enter Description">
           </b-form-input>
       </b-form-group>
@@ -70,18 +76,19 @@
           <b-form-input id="form-metadata-input"
                         type="text"
                         v-model="addFavoriteForm.metaData"
-                        required
-                        placeholder="Enter Metadata">
+                        placeholder="meta-data should be in key:value pair">
           </b-form-input>
       </b-form-group>
       <b-form-group id="form-category-group"
                     label="Category:"
                     label-for="form-category-input">
           <b-select v-model="addFavoriteForm.categoryId">
+          <option :value="null" selected="selected">select category</option>
           <option v-for="category in categories" v-bind:value="category.id">
             {{ category.type }}
           </option>
         </b-select>
+        <span class="category" v-b-modal.category-modal>Add Category</span>
       </b-form-group>
       <b-form-group id="form-rank-group"
                     label="Rank:"
@@ -159,13 +166,29 @@
       </b-button-group>
     </b-form>
   </b-modal>
+     <b-modal ref="addCategoryModal" id="category-modal" title="New Category" hide-footer>
+      <b-form @submit="onSubmitCategory" @reset="onReset" class="w-100">
+        <b-form-group id="form-type-group" label="Name:" label-for="form-type-input">
+          <b-form-input id="form-type-input" type="text" v-model="addCategoryForm.type" required placeholder="Enter category">
+          </b-form-input>
+        </b-form-group>
+
+        <b-button-group>
+          <b-button type="submit" variant="success">Create</b-button>
+          <b-button type="reset" variant="danger">Reset</b-button>
+        </b-button-group>
+      </b-form>
+    </b-modal>
   </div>
 </template>
 
 
 <script>
 import axios from 'axios';
+import moment from 'moment'
 import Alert from './Alert.vue';
+import Swal from 'sweetalert2';
+const BASE_URL = 'http://127.0.0.1:5000/api/v1'
 export default {
   // name: 'Favorite',
   data() {
@@ -186,6 +209,9 @@ export default {
         categoryId: '',
         rank:''
       },
+      addCategoryForm: {
+        type: ''
+      },
       message: '',
       showMessage: false,
       categories: [],
@@ -198,33 +224,58 @@ export default {
   },
   methods: {
     getFavorite(){
-      const path = 'http://127.0.0.1:5000/api/v1/favorites';
+      const path = `${BASE_URL}/favorites`;
        var headers = {
         'Authorization': `Bearer ${this.token}`
       }
-      // console.log('===>', headers)
       axios.get(path, {headers}).then((response) => {
-        console.log('===>>>>', response)
+         if(response.data.data.length === 0){
+            this.message = 'You don\'t have any favorite things yet'
+            this.showMessage = true
+          }
         this.favorites = response.data.data;
       })
         .catch((error) => {
-          console.log(error);
+          console.log(' i got error', error)
+
+          this.message = error
+          this.showMessage = true
         });
     },
 
      getCategories(){
-      const path = 'http://127.0.0.1:5000/api/v1/categories';
+      const path = `${BASE_URL}/categories`;
       axios.get(path).then((response) => {
-        console.log('=====>>>', response.data.data)
         this.categories = response.data.data;
       })
         .catch((error) => {
-          console.log(error);
+          this.message = error
         });
     },
 
+    addCategory(payload){
+      const path = `${BASE_URL}/categories`;
+       var headers = {
+        'Authorization': `Bearer ${this.token}`
+      }
+      axios.post(path, payload, {headers}).then((response) => {
+        console.log('====>>',response)
+        this.showMessage = true
+        this.message = response.data.message
+        this.$refs.addCategoryModal.hide()
+        this.getCategories()
+
+      })
+      .catch((error) => {
+        this.message = response.data.message
+        this.$refs.addCategoryModal.hide()
+
+      })
+
+    },
+
     addFavorite(payload){
-      const path = 'http://127.0.0.1:5000/api/v1/favorites';
+      const path = `${BASE_URL}/favorites`;
       var headers = {
         'Authorization': `Bearer ${this.token}`
       }
@@ -237,7 +288,7 @@ export default {
       })
       .catch((error) => {
         this.message = response.data.message
-        this.getBooks();
+        this.getFavorite();
       })
 
     },
@@ -246,16 +297,25 @@ export default {
       this.editFavoriteForm = favorite;
     },
     updateFavorite(payload, favorite_id){
-    const path = `http://127.0.0.1:5000/api/v1/favorite/${favorite_id}`;
-    axios.patch(path, payload)
+      var headers = {
+        'Authorization': `Bearer ${this.token}`
+      }
+    const path = `${BASE_URL}/favorites/${favorite_id}`;
+    axios.patch(path, payload, {headers})
       .then((response) => {
+        Swal.fire({
+          position: 'top-end',
+          type: 'success',
+          title: response.data.message,
+          showConfirmButton: false,
+          timer: 1500
+        })
         this.getFavorite();
         this.showMessage = true
         this.message = response.data.message
       })
       .catch((error) => {
         // eslint-disable-next-line
-        console.error(error);
         this.getFavorite();
       });
     },
@@ -268,7 +328,8 @@ export default {
          this.editFavoriteForm.title = '',
         this.editFavoriteForm.description = '',
         this.editFavoriteForm.category = '',
-        this.editFavoriteForm.metaData = ''
+        this.editFavoriteForm.metaData = '',
+        this.addCategoryForm.type = ''
     },
 
     onSubmit(evt){
@@ -298,7 +359,8 @@ export default {
         title: this.editFavoriteForm.title,
         description: this.editFavoriteForm.description,
         categoryId: this.editFavoriteForm.categoryId,
-        metaData: this.editFavoriteForm.metaData
+        metaData: this.editFavoriteForm.metaData,
+        rank: this.editFavoriteForm.rank
       }
       this.updateFavorite(payload, this.editFavoriteForm.id);
       this.initForm();
@@ -312,12 +374,23 @@ export default {
     },
 
     deleteFavorite(favorite_id) {
-      const path = `http://127.0.0.1:5000/api/v1/favorite/${favorite_id}`;
-      axios.delete(path)
+      const path = `${BASE_URL}/favorites/${favorite_id}`;
+      var headers = {
+        'Authorization': `Bearer ${this.token}`
+      }
+      axios.delete(path, {headers})
         .then((response) => {
-          this.getFavorite();
-          this.message = response.data.message;
-          this.showMessage = true;
+          if(response.data.status == 'success'){
+            Swal.fire(
+            'Deleted!',
+            'Favorite has been deleted.',
+            'success'
+            )
+            this.getFavorite();
+            this.message = response.data.message;
+            this.showMessage = true;
+          }
+
         })
         .catch((error) => {
           // eslint-disable-next-line
@@ -326,12 +399,53 @@ export default {
         });
     },
     onDeleteFavorite(favorite) {
-      this.deleteFavorite(favorite.id);
+      Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.deleteFavorite(favorite.id)
+      }
+    })
     },
+
+    onSubmitCategory(evt) {
+      evt.preventDefault();
+      this.$refs.addCategoryModal.hide()
+      const payload = {
+        type: this.addCategoryForm.type,
+      }
+      this.addCategory(payload);
+      this.initForm();
+
+    },
+    formatDate(date) {
+         return moment(String(date)).format('hh:mm a MM/D/YY');
+     }
   },
    created() {
     this.getFavorite();
     this.getCategories();
   },
+   updated() {
+  },
 };
 </script>
+<style scoped>
+  .category {
+    float: right;
+
+  }
+  .linkcontainer {
+    display: flex;
+    width: 200px;
+  }
+  .linkcontainer a {
+      flex: 1;
+    }
+</style>
